+++ lib/passive_liveness.dart
@@ -9,6 +9,7 @@ export 'src/passive_liveness_detector.dart';
 export 'src/ui/adaptive_screen_flash.dart';
 export 'src/utils/color_space_analyzer.dart';
 export 'src/utils/face_proximity_gate.dart';
+export 'src/utils/high_res_screen_analyzer.dart';
 export 'src/utils/image_preprocessor.dart';
 export 'src/utils/lbp_hog_analyzer.dart';
 export 'src/utils/liveness_logger.dart';